requirements.txt
pyautogui
Pillow
requests
python-dotenv
rich
pydantic
pyreadline3; sys_platform == 'win32'
